+++ lib/lite/ruby/enumerable.rb
@@ -14,6 +14,10 @@ module Enumerable
     end
   end
 
+  def cluster_by(&block)
+    group_by(&block).sort.transpose.pop || []
+  end
+
   def deduce(identity = 0, &block)
     if block_given?
       map(&block).deduce(identity)
@@ -72,6 +76,12 @@ module Enumerable
     end
   end
 
+  def frequency
+    each_with_object(Hash.new(0)) { |val, hash| hash[val] += 1 }
+  end
+
+  alias occurrences frequency
+
   # rubocop:disable Style/CaseEquality
   def incase?(object)
     any? { |val| object === val }
@@ -117,9 +127,42 @@ module Enumerable
     end
   end
 
-  def occurrences
-    each_with_object(Hash.new(0)) { |key, hsh| hsh[key] += 1 }
+  def modulate(modulo)
+    if modulo == 1
+      to_a
+    elsif size % modulo != 0
+      raise ArgumentError, "Invalid modulo: #{modulo.inspect}"
+    else
+      (0...size).each_with_object(Array.new(modulo, [])) do |i, array|
+        array[i % modulo] += [self[i]]
+      end
+    end
   end
+
+  # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+  def occur(amount = nil)
+    result = Hash.new { |hash, key| hash[key] = [] }
+
+    each do |item|
+      key = item
+      result[key] << item
+    end
+
+    if block_given?
+      result.select! { |_key, values| yield(values.size) }
+    else
+      raise ArgumentError, 'Invalid occur amount' unless amount
+
+      if amount.is_a?(Range)
+        result.select! { |_key, values| amount.include?(values.size) }
+      else
+        result.select! { |_key, values| values.size == amount }
+      end
+    end
+
+    result.values.flatten.uniq
+  end
+  # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
 
   def produce(identity = 0, &block)
     if block_given?
@@ -148,6 +191,23 @@ module Enumerable
 
     found_count > 1
   end
+
+  # rubocop:disable Metrics/MethodLength
+  def squeeze(*limited_to)
+    first = true
+    current = nil
+
+    each_with_object([]) do |val, array|
+      if !limited_to.empty? && !limited_to.include?(val)
+        array << val
+      elsif first || current != val
+        array << val
+        first = false
+        current = val
+      end
+    end
+  end
+  # rubocop:enable Metrics/MethodLength
 
   def take_last(num)
     collection_size = to_a.size