+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2019-07-18
 ### Added
+- Added Enumerable => `cluster_by`
+- Added Enumerable => `frequency`
+- Added Enumerable => `occur`
+- Added Enumerable => `squeeze`
 - Added Hash => `alias`
 - Added Hash => `collate`
 - Added Hash => `dearray_values`